+++ core/api/golang/lib/binding_constructors/binding_constructors.go
@@ -316,7 +316,7 @@ func NewWaitForHttpPostEndpointAvailabilityArgs(
 }
 
 // ==============================================================================================
-//                           Upload Files Artifact
+//                                  Upload Files Artifact
 // ==============================================================================================
 func NewUploadFilesArtifactArgs(data []byte) *kurtosis_core_rpc_api_bindings.UploadFilesArtifactArgs {
 	return &kurtosis_core_rpc_api_bindings.UploadFilesArtifactArgs{Data: data}
@@ -324,14 +324,14 @@ func NewUploadFilesArtifactArgs(data []byte) *kurtosis_core_rpc_api_bindings.Upl
 
 
 // ==============================================================================================
-//                           Store Web Files Artifact
+//                                 Store Web Files Artifact
 // ==============================================================================================
 func NewStoreWebFilesArtifactArgs(url string) *kurtosis_core_rpc_api_bindings.StoreWebFilesArtifactArgs {
 	return &kurtosis_core_rpc_api_bindings.StoreWebFilesArtifactArgs{Url: url}
 }
 
 // ==============================================================================================
-//                       Store Files Artifact From Service
+//                              Store Files Artifact From Service
 // ==============================================================================================
 func NewStoreFilesArtifactFromServiceArgs(serviceId string, sourcePath string) *kurtosis_core_rpc_api_bindings.StoreFilesArtifactFromServiceArgs {
 	return &kurtosis_core_rpc_api_bindings.StoreFilesArtifactFromServiceArgs{ServiceId: serviceId, SourcePath: sourcePath}